pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "azure-rag-llmops"
version = "0.1.0"
description = "RAG with Azure OpenAI + Azure Cognitive Search on FastAPI"
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
  "azure-identity==1.16.1",
  "azure-search-documents==11.6.0",        # stable
  "openai>=1.30.0",                        # new SDK
  "langchain==0.2.14",                     # 0.2.x
  "langchain-openai==0.1.22",              # OpenAI integrations moved here
  "langchain-community==0.2.12",           # AzureSearch vectorstore lives here
  "python-dotenv==1.0.1",
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "tiktoken==0.7.0",
  "prometheus-fastapi-instrumentator==6.1.0",
  "httpx==0.27.0",
  "pydantic-settings>=2.2",
  "pydantic>=2.6"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
  "types-requests",
]

[tool.pytest.ini_options]
addopts = "-q --cov=webapp --cov-report=term-missing"
pythonpath = ["webapp"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

# >>> This is the key fix <<<
[tool.hatch.build.targets.wheel]
packages = ["webapp"]

[tool.hatch.build.targets.sdist]
include = ["webapp", "README.md", "pyproject.toml", "k8s", ".github", "Dockerfile", ".env.example"]
